+++ src/llama-context.h
@@ -419,7 +419,4 @@ struct llama_context {
     void set_mtp_op_type(llama_mtp_op_type value);
 
     int max_nodes(int n_tokens, int n_kv) const;
-
-    int  graph_generation    = -1;   // last model generation we built for
-    bool force_graph_rebuild = false; // set true after any reload
 };
